Alpha 0
Just a simple function in R I made that makes coding data slightly
easier. Enter your dataframe, the name of the first data filled column
you would like to code as a string, the name of your first coding column
(must be made beforehand due to R being R, making the dataframe in Excel
might help), and then set auto_fill to either true or false. Auto fill
means that if you code an entry, all other entries that are the same
will receive the same coding. Note: Your data has to be all NAs for this
function to be useful, as it will skip already filled entries.